refactor(babel-preset-startupjs): rename patched metro preset
